+++ plugin/mutate/testdata/id_gen_algo/program.cpp
@@ -1,0 +1,11 @@
+/// @copyright Boost License 1.0, http://boost.org/LICENSE_1_0.txt
+/// @date 2022
+/// @author Joakim Brännström (joakim.brannstrom@gmx.com)
+
+#ifdef FOO
+
+bool fn1() { return true; }
+
+#endif
+
+bool fn2() { return true; }